+++ src/main/java/org/jabref/logic/ai/AiDefaultPreferences.java
@@ -10,7 +10,6 @@ import org.jabref.model.ai.EmbeddingModel;
 public class AiDefaultPreferences {
     public enum PredefinedChatModel {
         GPT_4O_MINI(AiProvider.OPEN_AI, "gpt-4o-mini", 128000),
-        GPT_LOCAL(AiProvider.GPT4ALL, "", 128000),
         GPT_4O(AiProvider.OPEN_AI, "gpt-4o", 128000),
         GPT_4(AiProvider.OPEN_AI, "gpt-4", 8192),
         GPT_4_TURBO(AiProvider.OPEN_AI, "gpt-4-turbo", 128000),
@@ -24,7 +23,8 @@ public class AiDefaultPreferences {
         GEMINI_1_5_PRO(AiProvider.GEMINI, "gemini-1.5-pro", 2097152),
         GEMINI_1_0_PRO(AiProvider.GEMINI, "gemini-1.0-pro", 32000),
         // Dummy variant for Hugging Face models.
-        HUGGING_FACE(AiProvider.HUGGING_FACE, "", 0);
+        // Blank entry used for cases where the model name is not specified.
+        BLANK(AiProvider.HUGGING_FACE, "", 0);
 
         private final AiProvider aiProvider;
         private final String name;
@@ -63,8 +63,8 @@ public class AiDefaultPreferences {
             AiProvider.OPEN_AI, PredefinedChatModel.GPT_4O_MINI,
             AiProvider.MISTRAL_AI, PredefinedChatModel.OPEN_MIXTRAL_8X22B,
             AiProvider.GEMINI, PredefinedChatModel.GEMINI_1_5_FLASH,
-            AiProvider.HUGGING_FACE, PredefinedChatModel.HUGGING_FACE,
-            AiProvider.GPT4ALL, PredefinedChatModel.GPT_LOCAL
+            AiProvider.HUGGING_FACE, PredefinedChatModel.BLANK,
+            AiProvider.GPT4ALL, PredefinedChatModel.BLANK
     );
 
     public static final boolean CUSTOMIZE_SETTINGS = false;