+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We added automatic browser extension install on Windows for Chrome and Edge. [#6076](https://github.com/JabRef/jabref/issues/6076)
 - We added a search bar for filtering keyboard shortcuts. [#11686](https://github.com/JabRef/jabref/issues/11686)
 - By double clicking on a local citation in the Citation Relations Tab you can now jump the linked entry. [#11955](https://github.com/JabRef/jabref/pull/11955)
-- We fixed an issue that when we use the local model from Gpt4All, the response from the model is cuttoff.. [#11870](https://github.com/JabRef/jabref/issues/11870)
+- We fixed an issue that when we use the local model from Gpt4All, the response from the model is cuttoff. [#11870](https://github.com/JabRef/jabref/issues/11870)
 
 ### Changed